+++ geolocator_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- Removed implicit request for permissions when getting a position.
+
 ## 1.1.0
 
 - Added the [AndroidOptions] class to the `lib` directory.